mirror and cp should print size in --json mode.
```sh
mc cp -r --json /usr/bin /tmp | jq .
...
...
{
  "status": "success",
  "source": "/usr/bin/gcm-viewer",
  "target": "/tmp/bin/gcm-viewer",
  "size": 87128,
  "totalCount": 2602,
  "totalSize": 677422730
}
```

This allows scripts to build progress around this data.

+++ cmd/session-v8.go
@@ -48,7 +48,7 @@ type sessionV8Header struct {
 	LastCopied         string            `json:"lastCopied"`
 	LastRemoved        string            `json:"lastRemoved"`
 	TotalBytes         int64             `json:"totalBytes"`
-	TotalObjects       int               `json:"totalObjects"`
+	TotalObjects       int64             `json:"totalObjects"`
 }
 
 // sessionMessage container for session messages
